requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
numpy==1.24.3
opencv-python-headless==4.5.3.56
torch==1.9.0
torchvision==0.10.0
Pillow==8.3.1
ultralytics==8.0.0
google-generativeai==0.3.0
python-dotenv==0.19.0
gunicorn==20.1.0 
